requirements.txt
fastapi
databases
sqlalchemy
asyncpg
python-dotenv
uvicorn